글수정, 생성 api 유틸함수 react-query로 대체
