+++ client/web/src/utils/hooks/mutation/useMutatePost.ts
@@ -1,9 +1,12 @@
 import { CreateCommunityPostDto } from '@truepoint/shared/dist/dto/communityBoard/createCommunityPost.dto';
 import { UpdateCommunityPostDto } from '@truepoint/shared/dist/dto/communityBoard/updateCommunityPost.dto';
 import { AxiosError } from 'axios';
+import { useSnackbar } from 'notistack';
 import {
-  useMutation, UseMutationResult,
+  useMutation, UseMutationResult, useQueryClient,
 } from 'react-query';
+import { useHistory } from 'react-router-dom';
+import ShowSnack from '../../../atoms/snackbar/ShowSnack';
 import axios from '../../axios';
 
 // 글 생성 요청
@@ -14,9 +17,29 @@ const createPost = async (createPostDto: CreateCommunityPostDto) => {
 
 export const useCreatePost = (): UseMutationResult<
 any, AxiosError, CreateCommunityPostDto
-> => useMutation<any, AxiosError, CreateCommunityPostDto>(
-  createPost,
-);
+> => {
+  const { enqueueSnackbar } = useSnackbar();
+  const history = useHistory();
+  const queryClient = useQueryClient();
+  return useMutation<any, AxiosError, CreateCommunityPostDto>(
+    createPost,
+    {
+      onSuccess: (data, createPostDto) => {
+        const { platform: platformIndex } = createPostDto;
+        const platform = ['afreeca', 'twitch', 'free'][platformIndex];
+        queryClient.invalidateQueries(['community', { platform }], { refetchInactive: true });
+        ShowSnack('글 작성 성공', 'info', enqueueSnackbar);
+        history.push({
+          pathname: '/community-board',
+        });
+      },
+      onError: (error) => {
+        console.error(error);
+        ShowSnack('오류가 발생했습니다. 잠시 후 다시 시도해주세요', 'error', enqueueSnackbar);
+      },
+    },
+  );
+};
 
 type UpdateProps = {
   postId: number,
@@ -30,6 +53,24 @@ const updatePost = async (props: UpdateProps) => {
 };
 export const useUpdatePost = (): UseMutationResult<
 any, AxiosError, UpdateProps
-> => useMutation<any, AxiosError, UpdateProps>(
-  updatePost,
-);
+> => {
+  const { enqueueSnackbar } = useSnackbar();
+  const history = useHistory();
+  const queryClient = useQueryClient();
+  return useMutation<any, AxiosError, UpdateProps>(
+    updatePost,
+    {
+      onSuccess: (data, props) => {
+        const { postId } = props;
+        queryClient.invalidateQueries(['post', postId]);
+        queryClient.invalidateQueries(['community'], { refetchInactive: true });
+        ShowSnack('글 수정 성공', 'info', enqueueSnackbar);
+        history.goBack();
+      },
+      onError: (error) => {
+        console.error(error);
+        ShowSnack('오류가 발생했습니다. 잠시 후 다시 시도해주세요', 'error', enqueueSnackbar);
+      },
+    },
+  );
+};
